fix?: Change DB name in migrate.sh
From my point of view this value was not correct privously. This is
suprizing for me because it should not work with the old value but it
does... Clearly a sign that I do not quite get the project/tooling but
it is my best esthimate that this should be the right value.